requirements.txt
vk_api==11.9.9
vktools==1.0.6
python-dotenv==1.0.0